refactor : image/exception 폴더 밑으로 이동
