+++ src/main/java/com/kakaotech/team14backend/common/FileUtils.java
@@ -5,7 +5,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
-import com.kakaotech.team14backend.exception.ImageIOException;
+import com.kakaotech.team14backend.image.exception.ImageIOException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
