+++ src/main/java/com/kakaotech/team14backend/common/GlobalExceptionHandler.java
@@ -1,24 +1,14 @@
 package com.kakaotech.team14backend.common;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kakaotech.team14backend.common.ApiResponse;
-import com.kakaotech.team14backend.common.ApiResponseGenerator;
-import java.util.Map;
-
 import com.kakaotech.team14backend.exception.ExtentionNotAllowedException;
-import com.kakaotech.team14backend.exception.ImageIOException;
+import com.kakaotech.team14backend.image.exception.ImageIOException;
 import com.kakaotech.team14backend.exception.LastPostIdParameterException;
 import com.kakaotech.team14backend.exception.MemberNotFoundException;
 import com.kakaotech.team14backend.exception.NotEnoughPointException;
 import com.kakaotech.team14backend.exception.SizeParameterException;
-import com.kakaotech.team14backend.post.exception.MaxLevelSizeException;
-import com.kakaotech.team14backend.post.exception.PostLevelOutOfRangeException;
-import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.client.HttpClientErrorException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
